pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-memory-system"
version = "0.1.0"
description = "Long-term memory MCP server for persistent AI assistants"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "motor>=3.3.0",
    "pymongo>=4.6.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "chromadb>=0.5.5",
    "sentence-transformers>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
